add group rating routes, rename rating controllers, fix templates
