+++ app/Http/Controllers/RatingTopCommentsController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+class RatingTopCommentsController extends Controller
+{
+    public function index()
+    {
+        $commentsRating = getCommentsRating();
+
+        return view('rating.comments', [
+            'commentsRating' => $commentsRating,
+        ]);
+    }
+}
