+++ resources/views/rating/menu.blade.php
@@ -1,8 +1,8 @@
 <div class="text-center mb-3">
-    <a class="btn btn-light" href="{{ route('ratings.index') }}">
+    <a class="btn btn-light" href="{{ route('top.index') }}">
         {{ __('rating.page_title') }}
     </a>
-    <a class="btn btn-light" href="{{ route('ratings.comments') }}">
+    <a class="btn btn-light" href="{{ route('comments_top.index') }}"/>
         {{ __('rating.page_title_comments') }}
     </a>
 </div>
